requirements.txt
google-genai==1.7.0
urllib3==1.26.18
pypdf==5.4.0
openai==1.79.0
pandas==2.2.3
openpyxl==3.1.5
python-docx==1.1.2